Lock the cache file when updating metadata
In the previous implementation, updating the cache file happened in the
following order:
1) File is locked;
2) Data is read;
3) File is unlocked;
4) Data is modified;
5) File is locked;
6) Data is written;
7) File is unlocked.

As you might imagine, it could happen that another process writes to the
file between reading (step 2) and writing (step 6). Since the cache
would simply overwrite the contents on a write, those changes would have
been nullified.

This caused problems with a large number of parallel uploads where the
offset would effectively been reverted (while the new data has been
appended), resulting in larger and corrupt files.

This new implementation rearranges the steps a bit to the following:
1) File is locked;
2) Data is read;
3) Data is modified;
4) Data is written;
5) File is unlocked.

Now, two processes can no longer interfere with one another and simply
have to wait. This means that the FileStore takes a small performance
hit, but given that the alternative is corrupt files, that is a
non-issue.

I did my best to keep the method signature of the FileStore as unchanged
as possible for backwards compatibility, but this does mean that there
are some unused (deprecated) methods in there now.

Signed-off by: Jord Nijhuis <jord@nijhuis.me>